-updating all requests to handle request counter functionality for any user but admin -reading users from data base

+++ src/main/java/com/edafa/demo/dao/CourseRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface CourseRepo extends JpaRepository<Course,Integer> {
+public interface CourseRepo extends JpaRepository<Course,Long> {
 }
