+++ src/main/java/com/edafa/demo/dao/StudentRepo.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import java.util.List;
 import java.util.Optional;
 
-public interface StudentRepo extends JpaRepository<Student,Integer> {
+public interface StudentRepo extends JpaRepository<Student,Long> {
 
     List findByNationality(String nat);
     Student getById(int id);
